+++ client/common/lib/elements/dynamic-table-form.ts
@@ -15,24 +15,7 @@ import { unsafeHTML } from 'lit/directives/unsafe-html.js'
 import { classMap } from 'lit/directives/class-map.js'
 import { LivechatElement } from './livechat'
 import { ptTr } from '../directives/translation'
-
-// This content comes from the file assets/images/plus-square.svg, from the Feather icons set https://feathericons.com/
-const AddSVG: string =
-  `<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24"
-  fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round"
-  stroke-linejoin="round" class="feather feather-plus-square">
-    <rect x="3" y="3" width="18" height="18" rx="2" ry="2"></rect>
-    <line x1="12" y1="8" x2="12" y2="16"></line><line x1="8" y1="12" x2="16" y2="12"></line>
-  </svg>`
-
-// This content comes from the file assets/images/x-square.svg, from the Feather icons set https://feathericons.com/
-const RemoveSVG: string =
-  `<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24"
-  fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round"
-  stroke-linejoin="round" class="feather feather-x-square">
-    <rect x="3" y="3" width="18" height="18" rx="2" ry="2"></rect>
-    <line x1="9" y1="9" x2="15" y2="15"></line><line x1="15" y1="9" x2="9" y2="15"></line>
-  </svg>`
+import { AddSVG, RemoveSVG } from '../buttons'
 
 type DynamicTableAcceptedTypes = number | string | boolean | Date | Array<number | string>