+++ server/main.ts
@@ -19,6 +19,7 @@ import { BotsCtl } from './lib/bots/ctl'
 import { ExternalAuthOIDC } from './lib/external-auth/oidc'
 import { migrateMUCAffiliations } from './lib/prosody/migration/migrateV10'
 import { Emojis } from './lib/emojis'
+import { LivechatProsodyAuth } from './lib/prosody/auth'
 import decache from 'decache'
 
 // FIXME: Peertube unregister don't have any parameter.
@@ -51,6 +52,7 @@ async function register (options: RegisterServerOptions): Promise<any> {
   await initSettings(options)
 
   await Emojis.initSingleton(options) // after settings, before routes
+  await LivechatProsodyAuth.initSingleton(options)
 
   await initCustomFields(options)
   await initRouters(options)
@@ -115,6 +117,7 @@ async function unregister (): Promise<any> {
   await BotConfiguration.destroySingleton()
   await ExternalAuthOIDC.destroySingletons()
   await Emojis.destroySingleton()
+  await LivechatProsodyAuth.destroySingleton()
 
   const module = __filename
   OPTIONS?.peertubeHelpers.logger.info(`Unloading module ${module}...`)